requirements.txt
face_recognition
opencv-python
pillow 
imgaug
faiss-cpu
mtcnn
fastapi 
uvicorn
tensorflow-cpu
python-multipart